+++ test/various.jl
@@ -82,8 +82,8 @@ function escape_times(partnum=500; printinfo=true)
         println("+ escapetime works for Particle and MagneticParticle")
         println("  and understands Doors.")
         println("+ The escape time is always finite.")
-        println("+ randominside() work for FiniteWall and gives i.c.")
-        println("  inside mushroom cap.")
+        println("+ randominside() works for FiniteWall and gives i.c.")
+        println("  inside mushroom cap!")
         println("+ Required time: $(round(time()-tim, 3)) sec.")
     end
     return